🚚api: move certificationReportRepository from lib to src
